+++ lib/extensions/mod.rs
@@ -1,6 +1,7 @@
 pub mod date;
 pub mod head;
 pub mod logger;
+pub mod path_normalizer;
 pub mod ratelimit;
 pub mod real_ip;
 pub mod redirect;