+++ lib/lib.rs
@@ -75,22 +75,24 @@ pub mod extensions {
     //! Includes helpful middleware like Serve Static, Rate Limit and Logger.
     //!
     //! ## All Feature
-    //! | Name                 | Description                                           |
-    //! | -------------------- | ----------------------------------------------------- |
-    //! | [`Date`]             | Add the Date header to responses. Required by HTTP.   |
-    //! | [`Head`]             | Add support for HTTP `HEAD` requests.                 |
-    //! | [`Logger`]           | Log incoming requests to the console / file.          |
-    //! | [`RateLimiter`]      | Limit how many requests can be handled from a source. |
-    //! | [`RealIp`]           | Get the real IP of a client through a reverse proxy   |
-    //! | [`RedirectResponse`] | Shorthands for HTTP redirects.                        |
-    //! | [`RequestId`]        | Add a Request-Id header to all requests.              |
-    //! | [`RouteShorthands`]  | Shorthands for defining routes (`server.get(...)`).   |
-    //! | [`ServeStatic`]      | Serve static files from a dir.                        |
-    //! | [`Trace`]            | Add support for the HTTP `TRACE` method.              |
+    //! | Name                 | Description                                                               |
+    //! | -------------------- | ------------------------------------------------------------------------- |
+    //! | [`Date`]             | Add the Date header to responses. Required by HTTP.                       |
+    //! | [`Head`]             | Add support for HTTP `HEAD` requests.                                     |
+    //! | [`Logger`]           | Log incoming requests to the console / file.                              |
+    //! | [`PathNormalizer`]   | Normalize paths to a common format without trailing or repeating slashes. |
+    //! | [`RateLimiter`]      | Limit how many requests can be handled from a source.                     |
+    //! | [`RealIp`]           | Get the real IP of a client through a reverse proxy                       |
+    //! | [`RedirectResponse`] | Shorthands for HTTP redirects.                                            |
+    //! | [`RequestId`]        | Add a Request-Id header to all requests.                                  |
+    //! | [`RouteShorthands`]  | Shorthands for defining routes (`server.get(...)`).                       |
+    //! | [`ServeStatic`]      | Serve static files from a dir.                                            |
+    //! | [`Trace`]            | Add support for the HTTP `TRACE` method.                                  |
     pub use crate::_extensions::{
         date::{self, Date},
         head::Head,
         logger::{self, Logger},
+        path_normalizer::PathNormalizer,
         ratelimit::RateLimiter,
         real_ip::RealIp,
         redirect::{self, RedirectResponse, RedirectType},